mds: handle min_compat_client type change
in previous change, MDSMap::min_compat_client was changed from int8_t to
ceph_release_t, i.e. uint8_t, and in
Server::update_required_client_features(), we check the
MDSMap::min_compat_client to see if it is greater than given version, so
a negative "-1" would overflow and be interpreted as a 255, hence will
be always greater than whatever version is compared. so we need to
bump up the encoding version, and

* normalize the number if it is -1
* ignore the number which is way too large.

Signed-off-by: Kefu Chai <kchai@redhat.com>