restful: Expose perf counters
We currently do not expose perf counters in the restful module but this
is a useful piece of information to be exposed by the restful module.

Since we already have a function to gather all the information about the
perf counters, we can make this easily accessible through the restful
API.

This also supports filtering by daemons through regexps.

Signed-off-by: Boris Ranto <branto@redhat.com>