mgr/dashboard: Watch for pool pg's increase and decrease
Now when creating or editing a pool the background task will watch
increasing and decreasing pg's. The executing task will also use a
progress to show the progress.

If the change was not executed through the dashboard, there is no task
to show, but the frontend will make sure to inform the user about the
change that is going on in order to stop the user from doing actions on
the changing pool.

Fixes: https://tracker.ceph.com/issues/39482
Signed-off-by: Stephan Müller <smueller@suse.com>