rgw: protect removing bucket w radosgw-admin under multisite & w --bypass-gc
Under multisite we depend on the clean-up done by gc to add the
metadata information to remove tail objects across sites. When
radosgw-admin is used to remove a bucket and the --bypass-gc option is
used, this process is circumvented, thereby leaving orphan tail
objects across synced sites.

Rather than prevent the user from doing this, we instead warn them and
require them to provide the --yes-i-really-mean-it option to carry out
the operation.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>