mgr/dashboard: Set changeDetection to Push on Table Component
This will limit the number of times angular checks for changes to only once
when a new data object is pushed to the table component.

It will also remove the event of rendering the table when we hover a row.

Fixes: http://tracker.ceph.com/issues/39944

Signed-off-by: Tiago Melo <tmelo@suse.com>
(cherry picked from commit 7867c700a634f91fb579c1fbabd5d38452628591)