rgw: Get rid of ObjectCache::set_ctx
Many classes in RGW use the zombie-object (or perhaps unborn object)
antipattern, where they are constructed but invalid until some method
is called to finish constructing them.

Since I'm using this one as a demonstration of how to make an object
an observer and fixing it was trivial, I decided to fix it so we could
call add_observer in the constructor, rather than dealing with doing
everything in set_ctx().

Signed-off-by: Adam C. Emerson <aemerson@redhat.com>