msg: Support callback when message is sent.
To support copy-less writes in librados c api, we need to ensure that all references of a message's buffer have been released. When a pg is moved, its message will be resent, resulting two references in the messenger layer. Objecter should wait for the two references to be released before it can call message's onfinish.

To support this, we add a mechanism to call a callback when a message is sent in messenger layer, which Objecter can use to maintain message's references and call message's onfinish only when all references have been released.

Signed-off-by: Li Yichao <liyichao.good@gmail.com>