rgw_file: unify allocations of request id
The fix by Tao Chen <chentao@umcloud.com> for construction of the
request id in RGWWriteRequests/RGWContinuedReq paths--just writes,
effectively--revealed that RGWRequest::id was being set to zero
in general, while a valid request id was being assigned to
req_state.

This implies that code which the RGWRequest::id of non-write requests
would still use an invalid value, so avoid that by initializing
RGWRequest::id in all RGWLibRequests, and propagating its value
to req_state.  This still works as intended for RGWWriteRequest.

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>