test_volume_client: rewrite test_put_object_versioned
The test succeeds if it receives a CommandFailedError exception from the
embedded python progam. This manner of testing success is unreliable
since the embedded python program contains multiple statements and the
exception may be raised by any of those. Rewrite the test to fix this
issue.

Note: Couldn't test if test_put_object_versioned() works fine with this
rewriting since Python 2 compatible libraries are not generated anymore
even when -DWITH_PYTHON2=ON was passed as an argument to do_cmake.sh.
However, this should not be a problem since later commit on the same PR
would make it compatible with Python 3.

Fixes: http://tracker.ceph.com/issues/39510
Signed-off-by: Rishabh Dave <ridave@redhat.com>