debian: build against libSSL1.0 since civetweb needs it
The version of civetweb in Luminous is incompatible with libSSL1.1,
which is the default libSSL in e.g. Ubuntu Bionic. This means that if
you build ceph on Bionic, you get a civetweb that links against
libssl1.1 that then fails to run, producing logs like

2019-02-01 14:09:45.651071 7f3dd9065000  0 civetweb: 0x564bbd31cf40: load_dll: libcrypto.so.1.1: cannot find CRYPTO_num_locks

It's fixed in newer civetwebs, but for Luminous the answer is to
specify build-dependency on libssl1.0; we also then have to build
against libcurl4-gnutls-dev since libcurl4-openssl-dev depends on
libssl1.1-dev. We use alternative dependencies so on Xenial we
continue to build against libcurl4-openssl-dev.

Fixes: https://tracker.ceph.com/issues/20696
Signed-off-by: Matthew Vernon <mv3@sanger.ac.uk>