mgr/dashboard: Subscribe to changes when RequiredIf is used
FormControls that use 'requiredIf' validators depend on the state of other
controls.
To prevent that some of the state changes are not tracked, we now automatically
subscribe to the prerequsited controls and update the validation of the parent
control.

Fixes: https://tracker.ceph.com/issues/40017

Signed-off-by: Tiago Melo <tmelo@suse.com>
(cherry picked from commit 86285b543b7d4f594599365a4e78c474d7ebb0ae)