journal: optimize object overflow detection
Previously to detect overflow we were sending journal append
requests until -EOVERFLOW is returned by osd. This means that we
had at least one waste (rejected) request per object set (though
there may be more if the number of in-flight appends is not
limited).

We can easily predict when the osd will start to return
-EOVERFLOW and avoid such additional requests.

Signed-off-by: Mykola Golub <mgolub@suse.com>