rgw: lc: check for valid placement target before processing transitions
Check that we transition to another placement target only if such a placement
target is valid in the zone. Also log the bucket and rule id in this case to
identify the LCRule causing the issue.

This still doesn't error the API at a put bucket lifecycle call

Fixes: https://tracker.ceph.com/issues/39632

Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>
(cherry picked from commit 2912bbf591ae1624007b3efabfa1dbfd5a137ca6)

Conflicts:
     src/rgw/rgw_lc.cc
- ldout(oc.cct, ...) instead of ldpp_dout(oc.dpp, ...)