ceph.in: do not preload libasan if it is not found
before this change, `asan_lib_path` could be `None` when we check it to
see if it ends with `NOTFOUND`. this happens if WITH_SEASTAR=OFF or
WITH_ASAN=OFF, as in that case, find_package(Sanitizers) is not called,
hence `ASAN_LIBRARY` won't be set.

in this change, libasan is only preloaded if

- (WITH_SEASTAR=ON and CMAKE_BUILD_TYPE=Debug and ASAN_LIBRARY is found)
  or
- (WITH_ASAN)

Signed-off-by: Kefu Chai <kchai@redhat.com>
(cherry picked from commit 538cf0f8e2bb959a2348d95db3d8162be762c351)