rgw: meta_be, bucket: meta be context can be strongly typed
When passing around the ctx for higher level non-generic functionality
it is better to have a strongly typed context. This is especially important
in the bucket entrypoint and bucket instance case where we can accidentally
confuse the two. There were cases where we created one and used it where
the other one was needed.

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>