cls/rgw: keep issuing bilog trim ops after reset
The CLSRGWIssueBILogTrim class is intended to keep submitting
cls_log_trim operations on each bucket index shard until they return
-ENODATA to signal there's no more to trim.

However, if the while loop in CLSRGWConcurrentIO::operator() gets to the
end of the shard list and collects all of its completions,
wait_for_completions() will return false and exit the loop. This is
always the case when num_shards = 1.

Fixes: http://tracker.ceph.com/issues/40187

Signed-off-by: Casey Bodley <cbodley@redhat.com>