mgr/progress: Add event when OSD marked in
Basically keeps track of affected PGs
similar to how funciton _osd_out works
but instead of tracking up+acting set
we only care about acting. Also made
a minor change on the function pg_update:
we consider pg to be complete when it
is active and clean and nothing else.
Removed the variable "unmoved" since
If the pg wasn't moved, we shouldn't
have an event for it in the first place.

Signed-off-by: Kamoltat (Junior) Sirivadhna <ksirivad@redhat.com>