mgr: Add get_rates_from_data to mgr_util.py
The function calculates the difference between the values
in a time series list.

The previous location of it was in the dashboard, that's why the
function has doc tests. In order to be able to test them as part of the
manager utilities a bit of tox magic was added.

Fixes: https://tracker.ceph.com/issues/40365
Signed-off-by: Stephan Müller <smueller@suse.com>
(cherry picked from commit 34976bae148aec013b53e1376b45a7af98e69b93)
Signed-off-by: Ernesto Puerta <epuertat@redhat.com>

Conflicts:
	src/pybind/mgr/dashboard/controllers/osd.py