osd: flush object's content to preserve original content
Let's assume that we have an object named foo = "abcdefgh" in metadata pool,
and two chunks in chunk pool, chunk_A = "1234" and chunk_B = "ABCD".
Then we do

$ rados -p mp set-chunk foo 0 4 --target-pool cp chunk_A 0 --with-reference

$ rados -p mp set-chunk foo 4 4 --target-pool cp chunk_B 0 --with-reference

If we then try to read foo, we get "1234ABCD" instead of "abcdefgh".
The data stored in the original foo is completely lost.

This commit gives an option that can avoid it.

$ rados -p mp set-chunk foo 0 4 --target-pool cp --with-reference
  --using-orig-content

Signed-off-by: Myoungwon Oh <omwmw@sk.com>
Signed-off-by: Tsung-Ju (Andy) Lii <usefulalgorithm@gmail.com>