Update the sizing advice to match internals
In mailingslists I found that users on several occasions used inefficient block.db sizes, either to low, so that the next level of RocksDB does not fit, or in excess of it, wasting space that could've been used better with device classes.

I hope this change might help others to avoid misconfigurations.

Signed-off-by: Dominik Hannen <d5k+h4n@posteo.de>