qa: remove hadoop qa suite
Reason: teuthology has removed Ubuntu 14.04 as a target.

Note: no one is trying to support Hadoop and it apparently (for Luminous) has
an Ubuntu 14.04 dependency. master still has the hadoop suite with the latest
Ubuntu but, again, no one is working on supporting this or testing it. I lack
time to check what else may have been done to get the Hadoop suite working on
Ubuntu 14.04.

Fixes: http://tracker.ceph.com/issues/40286
Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>