rgw: fix a bug lead to a 403 return code in chunked upload mode
When uploading data in chunked upload mode, there may be
a truncated header at the end of receiving buffer in some
extreme cases. This will cause the get_padding_last_aws4_chunk_encoded()
function to return ERR_SIGNATURE_NO_MATCH which lead to an
upload failure. Stash the incomplete header and splices it
at the head of the receive buffer in the next round of reading.

Signed-off-by: Yang Chen <ax.victor@outlook.com>