crimson/osd: make do_peering_event() reentrant
because PeeringState could emit more events when processing a given
event, and we want to use a fixed place to collect the messages to be
sent by the states when processing these events. that's why rctx is set
before processing the event, and is reset after done with it. also,
when we set the rctx, we assert to verify that there is no existing rctx
in the place, to make sure we always flush the messages, before starting
processing another event.

in classic OSD, events are queued in the ShardedOpWQ as instances of
OpQueueItem. but in crimson OSD, we don't have such a queue, so in
crimson-osd, we need to find a new way to enable PeeringState to emit
more events and to enable the caller to collect the messages sent by them
without overwriting the existing rctx set when processing previous
events.

in this change, `PG` queues up the pending events in
`PG::pending_peering_events`, and process them in the FIFO manner, the
messages to be sent by them are collected by the `PeeringCtx` passed in
by the initial `PG::do_peering_event()` call.

the `private` variant of `PG::do_peering_event()` method is updated to
queue the new event and process the event in batch.

Signed-off-by: Kefu Chai <kchai@redhat.com>