ceph-volume: manually specified osd id legality check
Sometimes we need to deploy osd concurrently, in which case you need to manually specify the {osd-id} in the Prepare step. I think the specified {osd-id} does not have to be in the `osd tree`. As long as we guarantee that the manually specified {osd-id} is available in the cluster.

Signed-off-by: ypdai <self19900924@gmail.com>