rgw: Introduce bucket expiry option on empty buckets
This patch enables the users to auto-delete an empty bucket after a
preset expiry period. A typical use-case for this feature is to remove
residual empty buckets that will remain after all its objects have
expired.

Bucket expiry can be specified using LCRule by making use of the
tri-state field Status which now accepts 'Purge_Enabled' in addition to
'Enabled' and 'Disabled'.

The semantics to set a bucket expiry rule are:
1. Prefix must be the NULL string ""
2. Status must be "Purge_Enabled"
3. Bucket will be deleted only after expiry period (calculated since its
   creation time) AND only after the bucket is empty (incomplete MPUs
   will not be auto-aborted).

Signed-off-by: Prasad Krishnan <prasad.krishnan@flipkart.com>