common/config_values: set seastar logging level as well
when the logging level of a subsys changes, update seastar accordingly

before this change, seastar logging levels of different logger are not
connected to ceph's logging levels of different subsystem.

after this change, they are connected. whenever a subsys's logging level
changes, the corresponding seastar logger is updated as well.

Signed-off-by: Kefu Chai <kchai@redhat.com>