osd/osd_types: let pg_log_t::approx_size() return uint64_t
because both `head.version` and `tail.version` are of `uint64_t`, there
is no need to cast the difference between them to `size_t` which could
be a 32-bit unsigned integer on 32bit platforms. also, by casting it to
`size_t` makes it a little bit more difficult to compare it with
`osd_pg_log_trim_max` option, which is of `uint64_t`, as `std::min()`
requires both its parameters to be the same type if we want to use
tempalte parameter type deduction.

Signed-off-by: Kefu Chai <kchai@redhat.com>