ceph-volume: fix deploy playbook for ansible~=2.6,<2.7
When testing the lumionus and mimic branch the import_role
ansible feature does not work as expected. Variables created by one
role is not available to the next role called when using
import_role. To fix this, we need to use the roles keyword instead.

This commit is only applies to luminous and mimic and should
only be applied there.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>