os/bluestore: it's better to erase spanning blob only once
when several extents use the same spanning blob, we reclaim the
blob if all Pextent are to be released, but we erase the spanning blob
from spanning blob map several times. Although it's not a error, I think
it's better to erase the spanning blob with the last Oldextent who's
blob empty flag will be set true.

Signed-off-by: Xiangyang Yu <penglaiyxy@gmail.com>