client: don't report any vxattrs to listxattr
The convention with kernel filesystems is to not report vxattrs when
listxattr is called. Doing this can throw a wrench to archiving tools
that will attempt to restore files with xattrs intact, only to find
that some of them can't be stored.

Remove the code that prints out vxattrs in listxattr. With this we
also don't need the "hidden" flag in the vxattr definitions.

Also fix up the existing testcases to account for the change in
behavior.

Fixes: https://tracker.ceph.com/issues/40965
Signed-off-by: Jeff Layton <jlayton@redhat.com>