osd: don't carry PGLSFilter between multiple ops in MOSDOp.
Currently result of executing multiple PG listing operations
in single MOSDOp – if PGLSFilter is being used – depends on
the order in the sequence. To exemplify:

  [ CEPH_OSD_OP_PGNLS, CEPH_OSD_OP_PGNLS_FILTER ]

may bring different result for CEPH_OSD_OP_PGNLS than

  [ CEPH_OSD_OP_PGNLS_FILTER, CEPH_OSD_OP_PGNLS ].

This happens because ..._PGNLS_FILTER sets up PGLSFilter which
is used also by plain ..._PGNLS.

The commit changes that. After the filter is set-up separately
for each interested operation in MOSDOp.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>