mds: update atime with newer one when Fwr caps are flushed
Currently, xfstest generic/192 fails when run under kcephfs. It
creates a file on cephfs, and then reads it after a delay to ensure
that the atime is updated. It then unmounts and remounts the fs
and checks it again. cephfs fails at this point because the atime
reverts to its original value.

Fix this by updating the atime not only when we're flushing Fx caps,
but also when flushing Fw or Fr as well, and the atime is newer than
the one currently set on the inode.

Fixes: https://tracker.ceph.com/issues/41192
Signed-off-by: Jeff Layton <jlayton@redhat.com>