mount.ceph: fork a child to handle fetching secret from keyring
When a secret is not specified via mount options, then mmap a shared
buffer and spawn a child process to search the keyring for it.

Add a new routine that will create a CephContext, slurp in the keyring
and search it for a particular secret.

If one is found, then strcpy into the buffer and have the child exit
with a zero exit status. If the parent sees the child exit with any
other exit status, then it ignores the contents of the shared buffer.

For safety reasons, the child drops all capabilities other than
CAP_DAC_READ_SEARCH (to ensure that it'll be able to read the keyring)
should one be found. To achieve this, we add a new dependency on
libcap-ng.

Fixes: https://tracker.ceph.com/issues/16656
Signed-off-by: Jeff Layton <jlayton@redhat.com>